Make host names more configurable and use new names
