+++ paf-mvp-demo-express/src/operator.ts
@@ -1,6 +1,7 @@
 import express from "express";
-import {operator, publicKeys} from "./config";
+import {operator} from "./config";
 import {addOperatorApi} from "@operator/operator-api";
+import {publicKeys} from "./public-keys";
 
 export const operatorApp = express();
