+++ paf-mvp-demo-express/src/publisher.ts
@@ -1,9 +1,13 @@
 import express from "express";
-import {protocol} from "./config";
+import {cdn, cmp, protocol} from "./config";
 
 export const publisherApp = express();
 
 publisherApp.get('/', (req, res) => {
     const view = 'publisher/index';
-    res.render(view, {protocol});
+    res.render(view, {
+        protocol,
+        cmpDomain: cmp.host,
+        cdnDomain: cdn.host
+    });
 });
