+++ paf-mvp-demo-express/src/views/publisher/index.hbs
@@ -2,10 +2,10 @@
 <html lang="html">
 <head>
     <!-- CMP -->
-    <script src="{{protocol}}://cmp.com/assets/cmp/cmp.js"></script>
+    <script src="{{protocol}}://{{cmpDomain}}/assets/cmp/cmp.js"></script>
 
     <!-- Prebid SSO integration -->
-    <script onload="CMP.cmpCheck()" src="{{protocol}}://cdn.pafdemo.com/assets/paf-lib.js"></script>
+    <script onload="CMP.cmpCheck()" src="{{protocol}}://{{cdnDomain}}/assets/paf-lib.js"></script>
 
     <link rel="icon" type="image/png" sizes="32x32" href="/assets/favicon.png">
     <meta name="viewport" content="width=device-width"/>
